+++ src/lmgrep/features.clj
@@ -1,3 +1,4 @@
 (ns lmgrep.features)
 
+;; excluded by default
 (def raudikko? (= "true" (System/getenv "LMGREP_FEATURE_RAUDIKKO")))