+++ src/lmgrep/lucene/analyzer.clj
@@ -1,54 +1,13 @@
 (ns lmgrep.lucene.analyzer
   (:require [clojure.string :as str]
-            [lmgrep.features :as features])
-  (:import (java.util HashMap Map ArrayList Locale)
-           (java.text Collator)
+            [lmgrep.features :as features]
+            [lmgrep.lucene.predefined-analyzers :as lucene.predefined])
+  (:import (java.util HashMap Map)
+           (java.io File)
+           (java.nio.file Path)
            (org.apache.lucene.analysis.custom CustomAnalyzer CustomAnalyzer$Builder)
            (org.apache.lucene.analysis.util TokenizerFactory TokenFilterFactory CharFilterFactory)
-           (org.apache.lucene.analysis.en EnglishAnalyzer LovinsSnowballStemTokenFilterFactory)
-           (org.apache.lucene.analysis Analyzer CharArraySet)
-           (org.apache.lucene.analysis.ar ArabicAnalyzer)
-           (org.apache.lucene.analysis.standard StandardAnalyzer ClassicAnalyzer UAX29URLEmailAnalyzer)
-           (org.apache.lucene.analysis.bg BulgarianAnalyzer)
-           (org.apache.lucene.analysis.bn BengaliAnalyzer)
-           (org.apache.lucene.analysis.br BrazilianAnalyzer)
-           (org.apache.lucene.analysis.ca CatalanAnalyzer)
-           (org.apache.lucene.analysis.core StopAnalyzer UnicodeWhitespaceAnalyzer KeywordAnalyzer SimpleAnalyzer WhitespaceAnalyzer)
-           (org.apache.lucene.analysis.cjk CJKAnalyzer)
-           (org.apache.lucene.analysis.ckb SoraniAnalyzer)
-           (org.apache.lucene.analysis.cz CzechAnalyzer)
-           (org.apache.lucene.analysis.da DanishAnalyzer DanishSnowballStemTokenFilterFactory)
-           (org.apache.lucene.analysis.de GermanAnalyzer)
-           (org.apache.lucene.analysis.el GreekAnalyzer)
-           (org.apache.lucene.analysis.es SpanishAnalyzer)
-           (org.apache.lucene.analysis.et EstonianAnalyzer EstonianSnowballStemTokenFilterFactory)
-           (org.apache.lucene.analysis.eu BasqueAnalyzer BasqueSnowballStemTokenFilterFactory)
-           (org.apache.lucene.analysis.fa PersianAnalyzer)
-           (org.apache.lucene.analysis.fi FinnishAnalyzer)
-           (org.apache.lucene.analysis.fr FrenchAnalyzer)
-           (org.apache.lucene.analysis.ga IrishAnalyzer IrishSnowballStemTokenFilterFactory)
-           (org.apache.lucene.analysis.gl GalicianAnalyzer)
-           (org.apache.lucene.analysis.hi HindiAnalyzer)
-           (org.apache.lucene.analysis.hu HungarianAnalyzer)
-           (org.apache.lucene.analysis.hy ArmenianAnalyzer)
-           (org.apache.lucene.analysis.id IndonesianAnalyzer)
-           (org.apache.lucene.analysis.it ItalianAnalyzer)
-           (org.apache.lucene.analysis.lt LithuanianAnalyzer LithuanianSnowballStemTokenFilterFactory)
-           (org.apache.lucene.analysis.lv LatvianAnalyzer)
-           (org.apache.lucene.analysis.no NorwegianAnalyzer)
-           (org.apache.lucene.analysis.pt PortugueseAnalyzer)
-           (org.apache.lucene.analysis.ro RomanianAnalyzer RomanianSnowballStemTokenFilterFactory)
-           (org.apache.lucene.analysis.ru RussianAnalyzer)
-           (org.apache.lucene.analysis.ru RussianAnalyzer)
-           (org.apache.lucene.analysis.sv SwedishAnalyzer)
-           (org.apache.lucene.analysis.th ThaiAnalyzer)
-           (org.apache.lucene.analysis.tr TurkishAnalyzer TurkishSnowballStemTokenFilterFactory)
-           (org.apache.lucene.analysis.pl PolishAnalyzer)
-           (org.apache.lucene.analysis.hy ArmenianSnowballStemTokenFilterFactory)
-           (org.apache.lucene.analysis.nl DutchAnalyzer DutchSnowballStemTokenFilterFactory)
-           (org.apache.lucene.analysis.ca CatalanSnowballStemTokenFilterFactory)
-           (org.apache.lucene.analysis.nl KPSnowballStemTokenFilterFactory)
-           (org.apache.lucene.collation CollationKeyAnalyzer)))
+           (org.apache.lucene.analysis Analyzer)))
 
 ; https://lucene.apache.org/core/8_8_0/analyzers-common/constant-values.html#org.apache.lucene.analysis.miscellaneous.WordDelimiterGraphFilter.GENERATE_WORD_PARTS
 
@@ -58,57 +17,6 @@
   "Normalizes analysis component name."
   [component-name]
   (str/lower-case component-name))
-
-(def predefined-analyzers
-  (reduce (fn [acc [analyzer-name analyzer]]
-            (assoc acc (namify (str/replace analyzer-name "Analyzer" "")) analyzer))
-          {}
-          {"ArabicAnalyzer"            (ArabicAnalyzer.)
-           "BulgarianAnalyzer"         (BulgarianAnalyzer.)
-           "BengaliAnalyzer"           (BengaliAnalyzer.)
-           "BrazilianAnalyzer"         (BrazilianAnalyzer.)
-           "CatalanAnalyzer"           (CatalanAnalyzer.)
-           "CJKAnalyzer"               (CJKAnalyzer.)
-           "SoraniAnalyzer"            (SoraniAnalyzer.)
-           "StopAnalyzer"              (StopAnalyzer. (CharArraySet. (ArrayList.) true))
-           "CzechAnalyzer"             (CzechAnalyzer.)
-           "DanishAnalyzer"            (DanishAnalyzer.)
-           "DutchAnalyzer"             (DutchAnalyzer.)
-           "GermanAnalyzer"            (GermanAnalyzer.)
-           "GreekAnalyzer"             (GreekAnalyzer.)
-           "SpanishAnalyzer"           (SpanishAnalyzer.)
-           "EstonianAnalyzer"          (EstonianAnalyzer.)
-           "BasqueAnalyzer"            (BasqueAnalyzer.)
-           "PersianAnalyzer"           (PersianAnalyzer.)
-           "FinnishAnalyzer"           (FinnishAnalyzer.)
-           "FrenchAnalyzer"            (FrenchAnalyzer.)
-           "IrishAnalyzer"             (IrishAnalyzer.)
-           "GalicianAnalyzer"          (GalicianAnalyzer.)
-           "HindiAnalyzer"             (HindiAnalyzer.)
-           "HungarianAnalyzer"         (HungarianAnalyzer.)
-           "ArmenianAnalyzer"          (ArmenianAnalyzer.)
-           "IndonesianAnalyzer"        (IndonesianAnalyzer.)
-           "ItalianAnalyzer"           (ItalianAnalyzer.)
-           "LithuanianAnalyzer"        (LithuanianAnalyzer.)
-           "LatvianAnalyzer"           (LatvianAnalyzer.)
-           "NorwegianAnalyzer"         (NorwegianAnalyzer.)
-           "PortugueseAnalyzer"        (PortugueseAnalyzer.)
-           "RomanianAnalyzer"          (RomanianAnalyzer.)
-           "RussianAnalyzer"           (RussianAnalyzer.)
-           "ClassicAnalyzer"           (ClassicAnalyzer.)
-           "UAX29URLEmailAnalyzer"     (UAX29URLEmailAnalyzer.)
-           "SwedishAnalyzer"           (SwedishAnalyzer.)
-           "ThaiAnalyzer"              (ThaiAnalyzer.)
-           "TurkishAnalyzer"           (TurkishAnalyzer.)
-           "EnglishAnalyzer"           (EnglishAnalyzer.)
-           ;; add 13 MB to the binary
-           "PolishAnalyzer"            (PolishAnalyzer.)
-           "StandardAnalyzer"          (StandardAnalyzer.)
-           "UnicodeWhitespaceAnalyzer" (UnicodeWhitespaceAnalyzer.)
-           "WhitespaceAnalyzer"        (WhitespaceAnalyzer.)
-           "KeywordAnalyzer"           (KeywordAnalyzer.)
-           "SimpleAnalyzer"            (SimpleAnalyzer.)
-           "CollationKeyAnalyzer"      (CollationKeyAnalyzer. (Collator/getInstance Locale/ENGLISH))}))
 
 (defn stringify [m]
   (reduce (fn [acc [k v]] (assoc acc (name k) (str v))) {} m))
@@ -124,71 +32,71 @@
           {} (CharFilterFactory/availableCharFilters)))
 
 (def default-token-filters
-  (dissoc (assoc (reduce (fn [acc ^String token-filter-name]
-                           (assoc acc (namify token-filter-name) (TokenFilterFactory/lookupClass token-filter-name)))
-                         {} (TokenFilterFactory/availableTokenFilters))
-            (namify "lithuanianSnowballStem") LithuanianSnowballStemTokenFilterFactory
-            (namify "armenianSnowballStem") ArmenianSnowballStemTokenFilterFactory
-            (namify "basqueSnowballStem") BasqueSnowballStemTokenFilterFactory
-            (namify "catalanSnowballStem") CatalanSnowballStemTokenFilterFactory
-            (namify "danishSnowballStem") DanishSnowballStemTokenFilterFactory
-            (namify "dutchSnowballStem") DutchSnowballStemTokenFilterFactory
-            (namify "basqueSnowballStem") EstonianSnowballStemTokenFilterFactory
-            (namify "irishSnowballStem") IrishSnowballStemTokenFilterFactory
-            (namify "kpSnowballStem") KPSnowballStemTokenFilterFactory
-            (namify "turkishSnowballStem") TurkishSnowballStemTokenFilterFactory
-            (namify "romanianSnowballStem") RomanianSnowballStemTokenFilterFactory
-            (namify "lovinsSnowballStem") LovinsSnowballStemTokenFilterFactory)
-          "synonym"                                        ; because deprecated and requires a patch
-          ))
+  (reduce (fn [acc ^String token-filter-name]
+            (assoc acc (namify token-filter-name) (TokenFilterFactory/lookupClass token-filter-name)))
+          {} (TokenFilterFactory/availableTokenFilters)))
 
 (def token-filter-name->class
   (cond-> default-token-filters
           features/raudikko? (assoc (namify "raudikko")
                                     (import 'org.apache.lucene.analysis.fi.RaudikkoTokenFilterFactory))))
 
+(def analyzers
+  (reduce (fn [acc [analyzer-name analyzer-class]]
+            (assoc acc (namify (str/replace analyzer-name "Analyzer" "")) analyzer-class))
+          {}
+          lucene.predefined/analyzers))
+
 (def DEFAULT_TOKENIZER_NAME "standard")
+
+(defn ^Path config-dir->path [config-dir]
+  (let [^String dir (or config-dir ".")]
+    (.toPath (File. dir))))
+
+(defn get-component-or-exception [factories name component-type]
+  (if-let [component (get factories (namify name))]
+    component
+    (throw
+      (Exception.
+        (format "%s '%s' is not available. Choose one of: %s"
+                component-type
+                name
+                (sort (keys factories)))))))
+
+(defn custom-analyzer
+  ([opts]
+   (custom-analyzer opts char-filter-name->class tokenizer-name->class token-filter-name->class))
+  ([{:keys [config-dir char-filters tokenizer token-filters]}
+    char-filter-factories tokenizer-factories token-filter-factories]
+   (let [^CustomAnalyzer$Builder builder (CustomAnalyzer/builder ^Path (config-dir->path config-dir))]
+     (.withTokenizer builder
+                     ^Class (get-component-or-exception tokenizer-factories
+                                                        (get tokenizer :name DEFAULT_TOKENIZER_NAME)
+                                                        "Tokenizer")
+                     ^Map (HashMap. ^Map (stringify (get tokenizer :args))))
+
+     (doseq [{:keys [name args]} char-filters]
+       (.addCharFilter builder
+                       ^Class (get-component-or-exception char-filter-factories name "Char filter")
+                       ^Map (HashMap. ^Map (stringify args))))
+
+     (doseq [{:keys [name args]} token-filters]
+       (.addTokenFilter builder
+                        ^Class (get-component-or-exception token-filter-factories name "Token filter")
+                        ^Map (HashMap. ^Map (stringify args))))
+
+     (.build builder))))
 
 (defn ^Analyzer create
   "Either fetches a predefined analyzer or creates one from the config."
-  [{:keys [char-filters tokenizer token-filters analyzer]}]
+  [{:keys [analyzer] :as opts}]
   (try
     (or
-      (when-let [n (get analyzer :name)]
-        (when (nil? (get predefined-analyzers (namify (str/replace n "Analyzer" ""))))
-          (throw (Exception. (format "Analyzer '%s' is not available. Choose one of: %s"
-                                     (get analyzer :name)
-                                     (sort (keys predefined-analyzers))))))
-        (get predefined-analyzers (namify (str/replace n "Analyzer" ""))))
-      (let [^CustomAnalyzer$Builder cab (CustomAnalyzer/builder)]
-        (when (nil? (get tokenizer-name->class (namify (get tokenizer :name DEFAULT_TOKENIZER_NAME))))
-          (throw (Exception. (format "Tokenizer '%s' is not available. Choose one of: %s"
-                                     (get tokenizer :name)
-                                     (sort (keys tokenizer-name->class))))))
-        (.withTokenizer cab
-                        ^Class
-                        (get tokenizer-name->class (namify (get tokenizer :name DEFAULT_TOKENIZER_NAME)))
-                        ^Map (HashMap. ^Map (stringify (get tokenizer :args))))
-
-        (doseq [char-filter char-filters]
-          (when (nil? (get char-filter-name->class (namify (get char-filter :name))))
-            (throw (Exception. (format "Char filter '%s' is not available. Choose one of: %s"
-                                       (get char-filter :name)
-                                       (sort (keys char-filter-name->class))))))
-          (.addCharFilter cab
-                          ^Class (get char-filter-name->class (namify (get char-filter :name)))
-                          ^Map (HashMap. ^Map (stringify (get char-filter :args)))))
-
-        (doseq [token-filter token-filters]
-          (when (nil? (get token-filter-name->class (namify (get token-filter :name))))
-            (throw (Exception. (format "Token Filter '%s' is not available. Choose one of: %s"
-                                       (get token-filter :name)
-                                       (sort (keys token-filter-name->class))))))
-          (.addTokenFilter cab
-                           ^Class (get token-filter-name->class (namify (get token-filter :name)))
-                           ^Map (HashMap. ^Map (stringify (get token-filter :args)))))
-
-        (.build cab)))
+      (when-let [analyzer-name (get analyzer :name)]
+        (get-component-or-exception analyzers
+                                    (namify (str/replace analyzer-name "Analyzer" ""))
+                                    "Analyzer"))
+      (custom-analyzer opts))
     (catch Exception e
       (when (System/getenv "DEBUG_MODE")
         (.printStackTrace e))